alerts/requirements.txt
flask
twilio
python-dotenv
requests
schedule
